+++ api/test/term.e2e-spec.ts
@@ -9,9 +9,8 @@ describe('TermController (e2e)', () => {
   let testingUser: TestingUser;
   let anotherUser: TestingUser;
   let testProject: TestingProject;
-  let termId: string;
 
-  beforeAll(async () => {
+  beforeEach(async () => {
     app = await createAndMigrateApp();
     testingUser = await signupTestUser(app);
     anotherUser = await signupTestUser(app, 'another-user@test.com');
@@ -29,11 +28,18 @@ describe('TermController (e2e)', () => {
       .expect(201)
       .expect(res => {
         expect(res.body.data).toHaveExactProperties(['id', 'value', 'date']);
-        termId = res.body.data.id;
       });
   });
 
   it('/api/v1/projects/:projectId/terms (GET) should find terms for project', async () => {
+    await request(app.getHttpServer())
+      .post(`/api/v1/projects/${testProject.id}/terms`)
+      .set('Authorization', `Bearer ${testingUser.accessToken}`)
+      .send({
+        value: 'term.one',
+      })
+      .expect(201);
+
     await request(app.getHttpServer())
       .get(`/api/v1/projects/${testProject.id}/terms`)
       .set('Authorization', `Bearer ${testingUser.accessToken}`)
@@ -44,7 +50,44 @@ describe('TermController (e2e)', () => {
       });
   });
 
+  it('/api/v1/projects/:projectId/terms (GET) should return terms in lexical order', async () => {
+    const input = ['app.login', 'should be last', 'app.logout', '2 goes second', 'app.exit', 'menu.start', 'a term', '1 goes first'];
+    const expected = ['1 goes first', '2 goes second', 'a term', 'app.exit', 'app.login', 'app.logout', 'menu.start', 'should be last'];
+
+    for (const term of input) {
+      await request(app.getHttpServer())
+        .post(`/api/v1/projects/${testProject.id}/terms`)
+        .set('Authorization', `Bearer ${testingUser.accessToken}`)
+        .send({
+          value: term,
+        })
+        .expect(201);
+    }
+
+    await request(app.getHttpServer())
+      .get(`/api/v1/projects/${testProject.id}/terms`)
+      .set('Authorization', `Bearer ${testingUser.accessToken}`)
+      .expect(200)
+      .expect(res => {
+        expect(res.body.data).toHaveLength(expected.length);
+        expect(res.body.data.map(t => t.value)).toEqual(expected);
+      });
+  });
+
   it('/api/v1/projects/:projectId/terms/:termId (PATCH) should update term by id', async () => {
+    let termId: string;
+
+    await request(app.getHttpServer())
+      .post(`/api/v1/projects/${testProject.id}/terms`)
+      .set('Authorization', `Bearer ${testingUser.accessToken}`)
+      .send({
+        value: 'term.one',
+      })
+      .expect(201)
+      .expect(res => {
+        termId = res.body.data.id;
+      });
+
     await request(app.getHttpServer())
       .patch(`/api/v1/projects/${testProject.id}/terms/${termId}`)
       .set('Authorization', `Bearer ${testingUser.accessToken}`)
@@ -59,7 +102,48 @@ describe('TermController (e2e)', () => {
       });
   });
 
+  it('/api/v1/projects/:projectId/terms/:termId (PATCH) should accept terms with utf-8 encoding', async () => {
+    let termId: string;
+
+    await request(app.getHttpServer())
+      .post(`/api/v1/projects/${testProject.id}/terms`)
+      .set('Authorization', `Bearer ${testingUser.accessToken}`)
+      .send({
+        value: 'term.one',
+      })
+      .expect(201)
+      .expect(res => {
+        termId = res.body.data.id;
+      });
+
+    await request(app.getHttpServer())
+      .patch(`/api/v1/projects/${testProject.id}/terms/${termId}`)
+      .set('Authorization', `Bearer ${testingUser.accessToken}`)
+      .send({
+        value: 'term.two őúüöá 😀👍🍉你好',
+      })
+      .expect(200)
+      .expect(res => {
+        expect(res.body.data).toHaveExactProperties(['id', 'value', 'date']);
+        expect(res.body.data.id).toEqual(termId);
+        expect(res.body.data.value).toEqual('term.two őúüöá 😀👍🍉你好');
+      });
+  });
+
   it('/api/v1/projects/:projectId/terms/:termId (DELETE) should delete term by id', async () => {
+    let termId: string;
+
+    await request(app.getHttpServer())
+      .post(`/api/v1/projects/${testProject.id}/terms`)
+      .set('Authorization', `Bearer ${testingUser.accessToken}`)
+      .send({
+        value: 'term.one',
+      })
+      .expect(201)
+      .expect(res => {
+        termId = res.body.data.id;
+      });
+
     await request(app.getHttpServer())
       .delete(`/api/v1/projects/${testProject.id}/terms/${termId}`)
       .set('Authorization', `Bearer ${testingUser.accessToken}`)
@@ -75,6 +159,19 @@ describe('TermController (e2e)', () => {
   });
 
   it('/api/v1/projects/:projectId/terms should not access terms resource if not authenticated', async () => {
+    let termId: string;
+
+    await request(app.getHttpServer())
+      .post(`/api/v1/projects/${testProject.id}/terms`)
+      .set('Authorization', `Bearer ${testingUser.accessToken}`)
+      .send({
+        value: 'term.one',
+      })
+      .expect(201)
+      .expect(res => {
+        termId = res.body.data.id;
+      });
+
     await request(app.getHttpServer())
       .post(`/api/v1/projects/${testProject.id}/terms`)
       .expect(401);
@@ -93,6 +190,19 @@ describe('TermController (e2e)', () => {
   });
 
   it('/api/v1/projects/:projectId/terms should not access terms resource if not authorized', async () => {
+    let termId: string;
+
+    await request(app.getHttpServer())
+      .post(`/api/v1/projects/${testProject.id}/terms`)
+      .set('Authorization', `Bearer ${testingUser.accessToken}`)
+      .send({
+        value: 'term.one',
+      })
+      .expect(201)
+      .expect(res => {
+        termId = res.body.data.id;
+      });
+
     await request(app.getHttpServer())
       .post(`/api/v1/projects/${testProject.id}/terms`)
       .set('Authorization', `Bearer ${anotherUser.accessToken}`)
@@ -120,7 +230,7 @@ describe('TermController (e2e)', () => {
       .expect(404);
   });
 
-  afterAll(async () => {
+  afterEach(async () => {
     await app.close();
   });
 });