+++ bin/check.sh
@@ -1,9 +1,10 @@
-#!/bin/sh
+#!/bin/bash
 
 set -e
 
-function pprint()
-{
+export TERM=xterm
+
+function pprint() {
   local s=("$@") b w
   for l in "${s[@]}"; do
     ((w<${#l})) && { b="$l"; w="${#l}"; }
@@ -26,6 +27,11 @@ bin/test.sh
 
 pprint "Applying code format"
 yarn fmt
+
+if ! [[ -z $(git status -s) ]]; then
+    echo "You have uncommited changes on git or you might need to apply formatting to your source code (yarn fmt)"
+    exit 1
+fi
 
 pprint "Linting API code"
 cd api && yarn lint